requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.25
django-elasticsearch-dsl==8.0
djangorestframework==3.15.1
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.14.0
gunicorn==22.0.0
kombu==5.3.7
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
pillow